Fix C+P bug; nudge send semantics slightly more in the late-bound direction
Fact: In OROM, send semantics forces a 'vtable' field on objects.
Thought: not late-bound enough. Ties the whole thing to the machine model more than necessary. An object should be able to decide how it receives messages. That is what a computer on a network does!
Hypothesis: replacing the vtable *field* with a general *procedure* i.e. code path for a per-object "receive message" behaviour will be better.
Result: obviously more flexible, but quickly pulled in the direction of "obtaining code path via message send to another object" anyway.
Stuck but unclear on what.